[Refactor] RateBox.jsx(#4)
